requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.7.4
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
numpy==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.4
